[REFACTOR] RankButton 삭제 후 search-page로 추출
