+++ src/pages/home/search-page.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from "react";
 import BackButton from "../../components/common/BackButton";
 import ElementButton from "../../components/common/ElementButton";
-import RankButton from "../../components/common/RankButton";
 import SearchBar from "../../components/common/SearchBar";
 import useDebounce from "../../hooks/use-debounce";
 import MealRecipeCard from "../../components/cards/MealRecipeCard";
@@ -18,6 +17,11 @@ import { useMyProfileStore } from "../../stores/use-my-profile-store";
 import { formatRankingTime } from "../../utils/date";
 import alertImage from "../../assets/images/alert-circle.png";
 import MealRecipeCardSkeleton from "../../components/skeltons/MealRecipeCardSkeleton";
+
+// RankButton Props
+import Up from "../..//assets/icons/up-icon.svg";
+import Down from "../../assets/icons/down-icon.svg";
+import Same from "../../assets/icons/same-icon.svg";
 
 const SearchingPage = () => {
   const [keyword, setKeyword] = useState("");
@@ -174,3 +178,27 @@ const SearchingPage = () => {
 };
 
 export default SearchingPage;
+
+interface RankButtonProps {
+  rank: number;
+  name: string;
+  change: "UP" | "DOWN" | "SAME";
+  onClick?: () => void;
+}
+
+const RankButton = ({ rank, name, change, onClick }: RankButtonProps) => {
+  return (
+    <div
+      className="flex justify-between items-center w-full cursor-pointer"
+      onClick={onClick}
+    >
+      <div className="flex gap-2 text-zinc-800 text-base font-normal leading-5 shrink-0">
+        <p className="font-semibold">{rank}</p>
+        <p className="">{name}</p>
+      </div>
+      {change === "UP" && <img src={Up} alt="인기 상승" />}
+      {change === "DOWN" && <img src={Down} alt="인기 하락" />}
+      {change === "SAME" && <img src={Same} alt="인기 유지" />}
+    </div>
+  );
+};
